chore: move evaluation pkg to go directory
