+++ evaluation/go/clause_evaluator.go
