+++ evaluation/go/clause_evaluator_test.go
