+++ evaluation/go/rule_evaluator.go
