+++ evaluation/go/segment_evaluator.go
