+++ evaluation/go/strategy_evaluator.go
