+++ pkg/api/api/api_test.go
@@ -25,7 +25,7 @@ import (
 	"testing"
 	"time"
 
-	evaluation "github.com/bucketeer-io/bucketeer/evaluation"
+	evaluation "github.com/bucketeer-io/bucketeer/evaluation/go"
 	environmentproto "github.com/bucketeer-io/bucketeer/proto/environment"
 
 	"github.com/stretchr/testify/assert"
